+++ lib/google_analytics_mailer.rb
@@ -26,6 +26,9 @@ module GoogleAnalyticsMailer
 
   end
 
+  # Allow usage also in controllers since they have the same structure of ActionMailer
+  alias_method :google_analytics_controller, :google_analytics_mailer
+
   # This module provides methods to deal with parameter merging and similar stuff
   module GoogleAnalytics
 
@@ -55,3 +58,6 @@ end
 
 # Add the class method to ActionMailer::Base
 ActionMailer::Base.send :extend, GoogleAnalyticsMailer
+
+# Allow usage also in application controller if required in dependencies
+ActionController::Base.send :extend, GoogleAnalyticsMailer if defined?(ActionController::Base)